TST: Moved match_grids tests to separate file.
